chore(typescript): converted to TypeScript
