--- packages/victory-cursor-container/src/index.js
@@ -1,5 +1,0 @@
-export {
-  cursorContainerMixin,
-  default as VictoryCursorContainer,
-} from "./victory-cursor-container";
-export { default as CursorHelpers } from "./cursor-helpers";
